Level 2 Final Commit 1
Features decals, changes and completion to the level layout, more naming of objects, guard paths, a fix to (hopefully) stop the out side of the teleporter teleporting back to near the in side, some scenic changes and even more boundaries in the dockyard. If no major issues are identified and you don't see a "Level 2 Final Commit 2", assume this to be the final version on GitHub for specifically level 2, the city level.